\name{AdaptR-package}
\alias{}
\docType{package}
\title{
\packageTitle{AdaptR}
}
\description{
\packageDescription{AdaptR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AdaptR}
\packageIndices{AdaptR}
}
\author{
\packageAuthor{AdaptR}

Maintainer: \packageMaintainer{AdaptR}
}
\references{
Literature or other references for background information
}
\keyword{ package }
\seealso{

}
\examples{
AdaptR("C:/Parameter_File.txt")
}
