% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_pkgs}
\alias{assert_pkgs}
\title{Check availability of required packages}
\usage{
assert_pkgs(pkgs = NULL)
}
\arguments{
\item{pkgs, }{character vector with name(s) of package(s) to check.}
}
\value{
\code{TRUE} if all packages available, otherwise execution is halted with
an error.
}
\description{
Used internally, helper to check if SUGGESTED packages are available. Will
halt execution if any of the queried packages are not available.
}
\keyword{internal}
