% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat0}
\alias{cat0}
\title{cat() with sep = ""}
\usage{
cat0(...)
}
\arguments{
\item{...}{strings to be concatenated and printed.}
}
\description{
Used internally. Passes everything on to \link{cat} but enforces \code{sep = ""}.
Relates to \link{cat} as \link{paste0} relates to \link{paste}.
}
\keyword{internal}
