% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompactGrids_run.R
\name{CompactGrids}
\alias{CompactGrids}
\title{Create compacted environmental grids for AdaptR}
\usage{
CompactGrids(Pfile = Parameter.File.Path)
}
\arguments{
\item{Pfile}{A file path to the parameter file}
}
\value{
A set of compacted environment files in the specified location
}
\description{
Create compacted environmental grids for AdaptR
}
\examples{
CompactGrids("C:/MyDocuments/CompactorParameterFile.txt")
}

