% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_funs.R
\name{cond_rescale_prob}
\alias{cond_rescale_prob}
\title{Conditionally rescale probabilities}
\usage{
cond_rescale_prob(cond, prob, up = FALSE, rescale_factor)
}
\arguments{
\item{cond}{single logical, whether to rescale the probability or not. If
\code{FALSE} the input probability is returned without rescaling (ignoring all
other arguments), if \code{TRUE} the input probability is rescaled.}

\item{prob}{probability to be rescaled (as a proportion, i.e., in \verb{[0-1]}).}

\item{up}{single logical, if \code{FALSE} (default), the probability provided will
be rescaled downwards (i.e., closer to \code{0}, as \code{prob / rescale_factor}), if
\code{TRUE}, the probability provided will be rescaled upwards (i.e., closer to
\code{1}, as \code{1 - (1 - prob) / rescale_factor}).}

\item{rescale_factor}{single numerical value, the factor to rescale by as
described above. The \code{\link[=run_trial]{run_trial()}} function defines this as initial number
of trial arms divided by currently active number of trial arms.}
}
\value{
numerical, \code{prob} rescaled or not depending on inputs values.
}
\description{
Used internally. This function conditionally rescales probabilities, used in
\code{\link[=run_trial]{run_trial()}} based on the information specified in \code{\link[=setup_trial]{setup_trial()}},
\code{\link[=setup_trial_binom]{setup_trial_binom()}} or \code{\link[=setup_trial_norm]{setup_trial_norm()}}. Used to rescale adaptation
rule probability thresholds for superiority and/or inferiority in trials
without a common \code{control} arm.
}
\keyword{internal}
