% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trials.R
\name{dispatch_trial_runs}
\alias{dispatch_trial_runs}
\title{Simulate single trial after setting seed}
\usage{
dispatch_trial_runs(is, trial_spec, seeds, sparse, cores, cl = NULL)
}
\arguments{
\item{is}{vector of integers, the simulation numbers/indices.}

\item{trial_spec}{trial specification as provided by \code{\link[=setup_trial]{setup_trial()}},
\code{\link[=setup_trial_binom]{setup_trial_binom()}} or \code{\link[=setup_trial_norm]{setup_trial_norm()}}.}

\item{sparse}{single logical, as described in \code{\link[=run_trial]{run_trial()}}; defaults to
\code{TRUE} when running multiple simulations, in which case only the data
necessary to summarise all simulations are saved for each simulation.
If \code{FALSE}, more detailed data for each simulation is saved, allowing more
detailed printing of individual trial results and plotting using
\code{\link[=plot_history]{plot_history()}} (\code{\link[=plot_status]{plot_status()}} does not require non-sparse results).}

\item{cores}{single integer; the number of cores to run the simulations on
using the \code{parallel} library. Defaults to use the \code{"mc.cores"} global
option if set (\verb{options(mc.cores = <number>)}) and \code{1} otherwise; may be
increased to run multiple simulations in parallel. Use
\code{\link[parallel:detectCores]{parallel::detectCores()}} to find the number of available cores.}

\item{cl}{\code{NULL} (default) for running sequentially, otherwise a \code{parallel}
cluster for parallel computation if \code{cores > 1}.}
}
\value{
Single trial simulation object, as described in \code{\link[=run_trial]{run_trial()}}.
}
\description{
Helper function to dispatch the running of several trials to \code{\link[=lapply]{lapply()}} or
\code{\link[parallel:clusterApply]{parallel::parLapply()}}, seeting seeds correctly if a \code{base_seed} was used
when calling \code{\link[=run_trials]{run_trials()}}. Used internally in calls by the \code{\link[=run_trials]{run_trials()}}
function.
}
\keyword{internal}
