% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dispersal_Neighbourhood.R
\name{Dispersal_Neighbourhood}
\alias{Dispersal_Neighbourhood}
\title{Generate a dispersal neighbourhood file for AdaptR}
\usage{
Dispersal_Neighbourhood(radius, type = "neg.power", params = c(1, 1),
  output.name, output.directory, dispersal.plot = "FALSE")
}
\arguments{
\item{radius}{the maximum radius of the dispersal neighbourhood, in grid cell units}

\item{type}{the type of dispersal probability function to apply ('neg.power','neg.exponential')}

\item{params}{a vector of parameters for the dispersal function}

\item{output.name}{name for the dispersal neighbourhood}

\item{output.directory}{output folder in which to save the dispersal neighbourhood file}

\item{dispersal.plot}{whether to display a plot of the resultant dispersal probability against distance}
}
\value{
A dispersal neighbourhood for AdaptR output file (.dna) written to the specified folder, with the file labelled by the output.name
}
\description{
Generate a dispersal neighbourhood file for AdaptR
}
\examples{
# Generate a kernel for the drosophila example
filepath.data <- system.file("extdata", package="AdaptR")
Dispersal_Neighbourhood(radius=5, 
                       type="neg.power", 
                       params=c(1,1), 
                       output.name="Dispersal_relfile_L1_K1_rad5",
                       output.directory=file.path(filepath.data,"species_inputs"),
                       dispersal.plot=TRUE)  

}

