% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{equivalent_funs}
\alias{equivalent_funs}
\title{Assert equivalent functions}
\usage{
equivalent_funs(fun1, fun2)
}
\arguments{
\item{fun1, fun2}{functions to compare.}
}
\value{
Single logical.
}
\description{
Used internally. Compares the definitions of two functions (ignoring
environments, bytecodes, etc., by only comparing function arguments and
bodies, using \code{\link[=deparse]{deparse()}}).
}
\keyword{internal}
