% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmt_dig}
\alias{fmt_dig}
\title{Format digits before printing}
\usage{
fmt_dig(x, dig)
}
\arguments{
\item{x}{numeric, the numeric value(s) to format.}

\item{dig}{single integer, the number of digits.}
}
\value{
Formatted character string.
}
\description{
Used internally.
}
\keyword{internal}
