% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ys_and_draws.R
\name{get_ys_norm}
\alias{get_ys_norm}
\title{Generate normally distributed continuous outcomes}
\usage{
get_ys_norm(arms, means, sds)
}
\arguments{
\item{arms}{character vector, \code{arms} as specified in \link{setup_trial_norm}.}

\item{means}{numeric vector, true \code{means} in all \code{arms} as specified in
\link{setup_trial_norm}.}

\item{sds}{numeric vector, true standard deviations (\code{sds}) in all \code{arms} as
specified in \link{setup_trial_norm}.}
}
\value{
A function which takes the argument \code{allocs} (a character vector
with the allocations) and returns a numeric vector of the same length with
the corresponding, randomly generated outcomes (from normal distributions).
}
\description{
Used internally. Function factory used to generate a function that generates
outcomes from normal distributions.
}
\keyword{internal}
