% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{make_x_scale}
\alias{make_x_scale}
\title{Make x-axis scale for history/status plots}
\usage{
make_x_scale(x_value)
}
\arguments{
\item{x_value}{single character string, determining whether the number of
adaptive analysis looks (\code{"look"}, default), the total cumulated number of
participants randomised (\code{"total n"}) or the total cumulated number of
participants with outcome data available at each adaptive analysis
(\code{"followed n"}) are plotted on the x-axis.}
}
\value{
An appropriate scale for the \code{ggplot2} plot x-axis according to
the value specified in \code{x_value}.
}
\description{
Used internally. Prepares the x-axis scale for history/status plots.
Requires the \code{ggplot2} package installed.
}
\keyword{internal}
