% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapSummariseReplicates.R
\name{map.summary}
\alias{map.summary}
\title{Generate summary maps from AdpatR outputs}
\usage{
map.summary(output.folder.path = getwd(), summary.name = "summary")
}
\arguments{
\item{output.folder.path}{The directory where the output files are stored}

\item{summary.name}{An output name for the sumary files created}
}
\description{
Generate summary maps from AdpatR outputs
}
\examples{
# Using data example provided for Drosophila jambulina

}

