% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRangeSingleRun.R
\name{plot.range.single.run}
\alias{plot.range.single.run}
\title{Generate summary plot of range dynamics from a single run of AdpatR}
\usage{
plot.range.single.run(output.folder.path = getwd(), run.name)
}
\arguments{
\item{output.folder.path}{The directory where the output files are stored}

\item{run.name}{The name for the AdaptR simulation run}
}
\value{
summary plots of the entire distribution over time
}
\description{
Generate summary plot of range dynamics from a single run of AdpatR
}
