% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale}
\alias{rescale}
\title{Rescale numeric vector to sum to 1}
\usage{
rescale(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Numeric vector, \code{x} rescaled to sum to a total of 1.
}
\description{
Used internally.
}
\keyword{internal}
