% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop0_warning0}
\alias{stop0_warning0}
\alias{stop0}
\alias{warning0}
\title{stop() and warning() with call. = FALSE}
\usage{
stop0(...)

warning0(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}
}
\description{
Used internally. Calls \code{\link[=stop]{stop()}} or \code{\link[=warning]{warning()}} but enforces \code{call. = FALSE},
to suppress the call from the error/warning.
}
\keyword{internal}
