% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_num}
\alias{summarise_num}
\title{Summarise numeric vector}
\usage{
summarise_num(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
A numeric vector with seven named elements: \code{mean}, \code{sd}, \code{median},
\code{p25}, \code{p75}, \code{p0}, and \code{p100} corresponding to the mean, standard
deviation, median, and 25-/75-/0-/100-percentiles.
}
\description{
Used internally, to summarise numeric vectors.
}
\keyword{internal}
