% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vapply_helpers}
\alias{vapply_helpers}
\alias{vapply_num}
\alias{vapply_int}
\alias{vapply_str}
\alias{vapply_lgl}
\title{vapply helpers}
\usage{
vapply_num(X, FUN, ...)

vapply_int(X, FUN, ...)

vapply_str(X, FUN, ...)

vapply_lgl(X, FUN, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{\link[base]{expression}}
    object.  Other objects (including classed objects) will be coerced
    by \code{base::\link[base]{as.list}}.}

\item{FUN}{the function to be applied to each element of \code{X}:
    see \sQuote{Details}.  In the case of functions like
    \code{+}, \code{\%*\%}, the function name must be backquoted or quoted.}

\item{...}{optional arguments to \code{FUN}.}
}
\description{
Used internally. Helpers for simplifying code invoking vapply().
}
\keyword{internal}
