% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_int}
\alias{verify_int}
\title{Verify input is single integer (potentially within range)}
\usage{
verify_int(x, min_value = -Inf, max_value = Inf, open = "no")
}
\arguments{
\item{x}{value to check.}

\item{min_value, max_value}{single integers (each), lower and upper bounds
between which \code{x} should lie.}

\item{open}{single character, determines whether \code{min_value} and \code{max_value}
are excluded or not. Valid values: \code{"no"} (= closed interval; \code{min_value}
and \code{max_value} included; default value), \code{"right"}, \code{"left"}, \verb{"yes"``` (= open interval, }min_value\code{and}max_value` excluded).}
}
\value{
Single logical.
}
\description{
Used internally.
}
\keyword{internal}
