\name{AdaptR-package}
\alias{}
\docType{package}
\title{
\packageTitle{AdaptR}
}
\description{
\packageDescription{AdaptR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AdaptR}
\packageIndices{AdaptR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{AdaptR}

Maintainer: \packageMaintainer{AdaptR}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
