% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat0}
\alias{cat0}
\title{cat() with sep = ""}
\usage{
cat0(...)
}
\arguments{
\item{...}{strings to be concatenated and printed.}
}
\description{
Used internally. Passes everything on to \code{\link[=cat]{cat()}} but enforces \code{sep = ""}.
Relates to \code{\link[=cat]{cat()}} as \code{\link[=paste0]{paste0()}} relates to \code{\link[=paste]{paste()}}.
}
\keyword{internal}
