% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompactGrids_run.R
\name{CompactGrids}
\alias{CompactGrids}
\title{Create compacted environmental grids for AdaptR}
\usage{
CompactGrids(compactor.parameter.file.name, ncols, nrows, n.env.vars,
  n.time.points, raw.env.grids.name.files, output.env.name.file)
}
\arguments{
\item{compactor.parameter.file.name}{A file path to a text file (.txt) to be created by this function, where the grid compactor parameter file will be saved}

\item{ncols}{The number of columns on the spatial grid}

\item{nrows}{The number of rows on the spatial grid}

\item{n.env.vars}{The number of environment variables being considered}

\item{n.time.points}{The number of time points for which compacted environment grids are to be created}

\item{raw.env.grids.name.files}{A list of length n.env.vars of filepaths to text files holding the filepaths of individual environment grids to be compacted.}

\item{output.env.name.file}{A filepath to a text file (.txt) holding the filepaths where the compacted environment grids will be saved.}
}
\value{
A set of compacted environment files in the specified location
}
\description{
Create compacted environmental grids for AdaptR
}
\examples{
# Using data example provided for Drosophila jambulina
filepath.data <- system.file("extdata", package="AdaptR")

# create text files to describe the file path to each input variable, and the output variable
write.table((file.path(filepath.data,"Tmax",paste0("Tmax",seq(1:6),".asc"))), file = file.path(filepath.data,"Tmax","Tmax_filenames.txt"), eol = "\\n", row.names = FALSE, col.names = FALSE, quote=FALSE )
write.table((file.path(filepath.data,"Habitat",paste0("Habitat",seq(1:6),".asc"))), file = file.path(filepath.data,"Habitat","Habitat_filenames.txt"), eol = "\\n", row.names = FALSE, col.names = FALSE, quote=FALSE )
write.table((file.path(filepath.data,"compact_grids",paste0("demo_compact_grids_T",seq(1:6)))), file = file.path(filepath.data,"compact_grids","demo_compact_grids_output_filenames.txt"), eol = "\\n", row.names = FALSE, col.names = FALSE, quote=FALSE )

# Run the grid compactor
CompactGrids(compactor.parameter.file.name = file.path(filepath.data,"species_inputs","Jambulina__grid_compactor_parameter_file.txt"),
             ncols = 100,
             nrows = 79,
             n.env.vars = 2,
             n.time.points = 6,
             raw.env.grids.name.files = c(file.path(filepath.data,"Tmax","Tmax_filenames.txt"),file.path(filepath.data,"Habitat","Habitat_filenames.txt")),
             output.env.name.file = file.path(filepath.data,"compact_grids","demo_compact_grids_output_filenames.txt"))
}

