% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmt_pct}
\alias{fmt_pct}
\title{Create formatted label with absolute and relative frequencies (percentages)}
\usage{
fmt_pct(e, n, dec = 1)
}
\arguments{
\item{e}{integer, the numerator (e.g., the number of events).}

\item{n}{integer, the denominator (e.g., the total number of patients).}

\item{dec}{integer, the number of decimals for the percentage.}
}
\value{
Formatted character string.
}
\description{
Used internally.
}
\keyword{internal}
