% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapSingleRun.R
\name{map.single.run}
\alias{map.single.run}
\title{Generate summary maps from a single run of AdpatR}
\usage{
map.single.run(output.folder.path = getwd(), run.name)
}
\arguments{
\item{output.folder.path}{The directory where the output files are stored}

\item{run.name}{The name for the AdaptR simulation run}
}
\value{
A set of output files written in the specified folder, each containing the run.name
}
\description{
Generate summary maps from a single run of AdpatR
}
