% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics_ecdf.R
\name{plot_metrics_ecdf}
\alias{plot_metrics_ecdf}
\title{Plot empirical cumulative distribution functions of performance metrics}
\usage{
plot_metrics_ecdf(
  object,
  metrics = c("size", "sum_ys", "ratio_ys"),
  restrict = NULL,
  nrow = NULL,
  ncol = NULL,
  cores = getOption("mc.cores", 1)
)
}
\arguments{
\item{object}{\code{trial_results} object, output from the \code{\link[=run_trials]{run_trials()}}
function.}

\item{metrics}{the performance metrics to plot, as described in
\code{\link[=extract_results]{extract_results()}}. Multiple metrics may be plotted at the same time.
Valid metrics include: \code{size}, \code{sum_ys}, and \code{ratio_ys_mean}. All may be
specified with either spaces or underlines. Defaults to plotting all three.}

\item{restrict}{single character string or \code{NULL}. If \code{NULL} (default),
results are summarised for all simulations; if \verb{"superior}, results are
summarised for simulations ending with superiority only; if \code{"selected"},
results are summarised for simulations ending with a selected arm
(according to the specified arm selection strategy for simulations not
ending with superiority). Some summary measures (e.g., \code{prob_conclusive})
have substantially different interpretations if restricted, but are
calculated nonetheless.}

\item{nrow, ncol}{the number of rows and columns when plotting multiple
metrics in the same plot (using faceting in \code{ggplot2}). Defaults to \code{NULL},
in which case this will be determined automatically.}

\item{cores}{single integer; the number of cores used to extract simulation
results using the \code{parallel} library. Defaults to use the \code{"mc.cores"}
global option if set (\verb{options(mc.cores = <number>)}) and \code{1} otherwise;
more cores may be used to extract large simulation results quicker. Use
\code{parallel::detectCores()} to find the number of available cores.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Plots empirical cumulative distribution functions (ECDFs) of numerical
performance metrics across multiple simulations from a \code{"trial_results"}
object returned by \code{\link[=run_trials]{run_trials()}}. Requires the \code{ggplot2} package installed.
}
\examples{
#### Only run examples if ggplot2 is installed ####
if (requireNamespace("ggplot2", quietly = TRUE)){

  # Setup a trial specification
  binom_trial <- setup_trial_binom(arms = c("A", "B", "C", "D"),
                                   control = "A",
                                   true_ys = c(0.20, 0.18, 0.22, 0.24),
                                   data_looks = 1:20 * 100)

  # Run multiple simulation with a fixed random base seed
  res_mult <- run_trials(binom_trial, n_rep = 25, base_seed = 678)

  # NOTE: the number of simulations in this example is smaller than
  # recommended - the plots reflect that, and show that performance metrics
  # are not stable and have likely not converged yet

  # Plot ECDFs of continuous performance metrics
  plot_metrics_ecdf(res_mult)

}

}
\seealso{
\code{\link[=check_performance]{check_performance()}}, \code{\link[=summary]{summary()}}, \code{\link[=extract_results]{extract_results()}},
\code{\link[=plot_convergence]{plot_convergence()}}.
}
