% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_null}
\alias{replace_null}
\alias{\%||\%}
\title{Replace NULL with other value (NULL-OR-operator)}
\usage{
a \%||\% b
}
\arguments{
\item{a, b}{atomic values of any type.}
}
\value{
If \code{a} is \code{NULL}, \code{b} is returned. Otherwise \code{a} is returned.
}
\description{
Used internally, primarily when working with list arguments, because, e.g.,
\code{list_name$element_name} yields \code{NULL} when unspecified.
}
\keyword{internal}
